Update PaintPanel.java
Added the copyShape and pasteShape methods. copyShape changes a flag used by mouseListener to indicate that the user will be copying the selected shape. A new shape is created and acts like an "add" shape, meaning the undo method will also work for copied shapes. Commented out changesList and shapeList as they were used for debug. Can be removed completely.